refactor(runtime): extract persisted tool-result normalization from claude_native
Pure move, no behaviour change. The logic that turns a persisted tool-result
string into replay-safe Anthropic content blocks — and decides whether an inline
image payload is plausible enough to replay — was 336 lines inside the
5,837-line claude_native.py, where it is neither Claude-specific nor easy to
find. It now lives in omnigent/runtime/tool_result_content.py with one
responsibility, alongside the other "stored history -> model input" modules.

Moved 14 symbols, dropping `_claude_` prefixes and making the five names
claude_native actually consumes public: tool_result_content_blocks,
strip_unparseable_image_output, blocks_from_parsed_list,
image_payloads_in_blocks, RehydratedContent. The store shapes it reads (MCP
ImageContent, the newline-joined form, the "Error: " prefix) are why it belongs
outside a provider adapter.

`_image_omitted_placeholder` moves out of runtime/prompt.py as public
image_omitted_placeholder, and prompt.py imports it. That removes
claude_native's reach into another module's private symbol -- it was the only
external consumer -- and lets the import be eager, since the new module is a
leaf. Placeholder text is unchanged.

Kept in claude_native.py, because they encode Claude Code's transcript and
metadata format rather than store shapes: _tool_use_result_for_content,
_tool_use_result_payload_omitted, _json_safe_tool_use_result,
_sanitize_cloned_tool_result_record, _repair_cloned_tool_use_result,
_redact_binary_blocks.

Dependency direction is one-way (claude_native -> runtime.tool_result_content)
with no cycle: the new module imports stdlib plus omnigent.json_types only, and
importing it alone pulls in no harness module. It carries a 9-line private copy
of the `_json_object` narrowing helper rather than importing it back out of
claude_native, which would invert the direction.

Tests split by ownership: five tests that exercise only moved symbols go to
tests/runtime/test_tool_result_content.py; transcript, clone and metadata tests
stay. The shared image fixtures move to tests/_image_fixtures.py, matching the
existing tests/_model_pools.py pattern, so neither module duplicates them.

Verified: 17 representative records (transcript build, strip, and clone
sanitation) are byte-identical before and after, modulo the wall-clock timestamp
field. 238 tests before, 238 after (233 + 5). 700 passed across the claude-native,
prompt, compaction, MCP, bridge, launcher and content-helper suites. ruff clean,
mypy clean, pyrefly 0 errors when run with the worktree SDK on its search path.

runtime/compaction.py is deliberately untouched; its redaction gap is a separate
bounded step.
